backend/pyproject.toml
[project]
name = "videohosting-backend"
version = "0.1.0"
description = "Video hosting backend with Clean Architecture"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "dishka>=1.3.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "aiokafka>=0.11.0",
    "faststream[kafka]>=0.5.0",
    "aioboto3>=13.3.0",
    "types-aiobotocore-s3>=2.13.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt[crypto]>=2.9.0",
    "python-multipart>=0.0.17",
    "aiosmtplib>=3.0.2",
    "psycopg[binary]>=3.3.2",
    "structlog>=25.5.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]
[dependency-groups]
dev = [
    "ruff>=0.15.2",
    "ty>=0.0.18",
]

[tool.ruff]
line-length = 120

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    ".venv",
    "**/migrations/**",
]

[tool.ruff.lint]
select = [
    "ALL"
]
ignore = [
    "D103",
    "D102",
    "D100",
    "D101",
    "D104",
    "D107",
    "D105",
    "ANN401",
    "ANN204",
    "ANN202",
    "ANN201",
    "ANN003",
    "ANN002",
    "N815",
    "TC001",
    "D203",
    "D212",
    "PLR2004",
    "TRY301",
    "RUF006",
    "PGH003",
    "RUF002",
    "D415",
    "FAST002",
    "FBT001",
    "FBT002",
    "FBT003",
    "EM101",
    "EM102",
]


[tool.ruff.lint.isort]
known-first-party = [
    "application",
    "controllers",
    "domain",
    "infrastructure",
    "config",
    "ioc",
    "main",
]


[tool.ty.environment]
python-version = "3.12"
root = ["src"]

[tool.ty.src]
exclude = ["**/migrations/**"]
